+++ NEWS.md
@@ -1,3 +1,9 @@
+# gris 0.07
+
+* replace llh2xyz with geocentric, driving proj4 'geocent'
+
+* removed quadmeshFromRaster (see quadmesh package on CRAN)
+
 # gris 0.0.6.9001
 
 * Added a `NEWS.md` file to track changes to the package.